Add files via upload
With growing demands and competitions in the market, using the Superstore Giant dataset I will seek to gain a deeper understanding of its operations. The goal of this project is to analyse historical data to answer critical questions and provide actionable insights to drive business decisions. I will focus on identifying the most profitable products, regions, and customer segments to target while avoiding underperforming areas.